Unify component creation
Previously, we had in place three different methods for inserting views
or components into the render hierarchy. Top-level views went through
a legacy path that used RenderBuffer. Outlet’s went through our
abandoned ShadowRoot abstraction. The `{{view}}` helper and non-top-
level components went through ComponentNode.

While this commit started as an attempt to implement `{{link-to}}`, it
ended up being an exercise in unifying the above paths to just use
ComponentNode, which is itself simplified in this commit.

In order to support LinkView, we also implemented the view’s
`classNameBindings` syntax. In our approach, we dynamically translate
the `classNameBindings` microsyntax into the HTMLBars statements, then
use those to build a template that represents the view’s wrapping
element at runtime.

Once we had the class name bindings working, we were able to rewrite
the path for top-level views to use ComponentNodes. This allowed us to
remove large amounts of legacy view code, and there is almost assuredly
more code that we can now audit and remove entirely.

* Added registerKeyword facility
* Implemented classNameBindings
* Added `link` hook to keywords
* Rewrote `{{outlet}}` to use ComponentNode
* Added utility for building component templates (with nested blocks)
* Changed hook invocation order (`willRender` comes after
  `willReceiveAttrs`)
* Top-level views now render through ComponentNode rather than legacy
  RenderBuffer code path.
* Added zipHash stream utility
* Implemented largely working `LinkView`/`{{link-to}}`