[BUGFIX beta] Prevent `classNames` from being duplicated.
When a component is created we pass all `attrs` as root properties. In
this case, that means that the `classNames` concatenated properties have
already been merged/handled.

The fix is to remove the processing of `attrs.classNames` from
`normalizeClass`, which prevents any classes provided via `classNames`
from being added twice.